Revert "Initial update to r6343"
This reverts commit e5a4dd23c6b83f60294f892227c6749610c43338.

+++ source/plugins/emailnot/md5.c
@@ -38,7 +38,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned int UINT4;
+typedef unsigned long int UINT4;
